Moves store component and related files into its own folder
