+++ App.tsx
@@ -14,9 +14,9 @@ import Config from 'react-native-config'
 import SplashScreen from 'react-native-splash-screen'
 import Toast from 'react-native-toast-message'
 
-import StoreProvider from './App/Store'
 import { initStoredLanguage } from './App/localization'
 import RootStack from './App/navigators/RootStack'
+import StoreProvider from './App/store/Store'
 import indyLedgers from './configs/ledgers/indy'
 import toastConfig from './configs/toast/toastConfig'
