+++ App/navigators/RootStack.tsx
@@ -3,17 +3,17 @@ import { createStackNavigator } from '@react-navigation/stack'
 import React, { useContext, useState } from 'react'
 import { Text, TouchableOpacity, View } from 'react-native'
 
-import { DispatchAction } from '../Reducer'
-import { Context } from '../Store'
 import { Colors } from '../Theme'
 import Arrow from '../assets/img/large-arrow.svg'
 import { Screens } from '../constants'
+import { DispatchAction } from '../store/reducer'
 import Onboarding from '../screens/Onboarding'
 import { pages, carousel } from '../screens/OnboardingPages'
 import PinCreate from '../screens/PinCreate'
 import PinEnter from '../screens/PinEnter'
 import Splash from '../screens/Splash'
 import Terms from '../screens/Terms'
+import { Context } from '../store/Store'
 
 import ScanStack from './ScanStack'
 import TabStack from './TabStack'
